requirements.txt
alembic==1.3.0
astroid==2.3.3
Babel==2.7.0
blinker==1.4
cffi==1.13.2
Click==7.0
colorama==0.4.1
cryptography==2.8
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-paginate==0.5.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gitdb2==2.0.6
GitPython==3.0.4
iso3166==1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.6.1
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pbr==5.4.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
schwifty==2018.9.1
six==1.13.0
smmap2==2.0.5
SQLAlchemy==1.3.10
stevedore==1.31.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
